+++ src/predicates/mod.rs
@@ -1,4 +1,6 @@
 mod expansion;
+mod orient2d;
 mod sign;
 
+pub use orient2d::orient2d;
 pub use sign::{Orientation, Sign};